# src/main/java/com/example/librarymanagementbackend/dto/authen/request/AuthenticationRequest.java
package com.example.librarymanagementbackend.dto.authen.request;

import lombok.*;
import lombok.experimental.FieldDefaults;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
@FieldDefaults(level = AccessLevel.PRIVATE)
public class AuthenticationRequest {
    String name;
    String password;
}