# src/main/java/com/example/librarymanagementbackend/dto/authen/request/LogoutRequest.java
package com.example.librarymanagementbackend.dto.authen.request;

import lombok.*;
import lombok.experimental.FieldDefaults;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
@FieldDefaults(level = AccessLevel.PRIVATE)
public class LogoutRequest {
    String refreshToken;
}