# src/main/java/com/example/librarymanagementbackend/dto/authen/request/RefreshRequest.java
package com.example.librarymanagementbackend.dto.authen.request;

import lombok.*;
import lombok.experimental.FieldDefaults;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
@FieldDefaults(level = AccessLevel.PRIVATE)
public class RefreshRequest {
    String refreshToken;
}