# swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Library Management API",
        "description": "Library Management System API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Development Server"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/auth/Refresh": {
            "post": {
                "tags": [
                    "authentication-controller"
                ],
                "operationId": "refresh",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAuthenticationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/Logout": {
            "post": {
                "tags": [
                    "authentication-controller"
                ],
                "operationId": "logout",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LogoutRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/Login": {
            "post": {
                "tags": [
                    "authentication-controller"
                ],
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthenticationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAuthenticationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "result": {
                        "type": "object"
                    },
                    "error": {
                        "type": "object"
                    }
                }
            },
...
            "RefreshRequest": {
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string"
                    }
                }
            },
            "ApiResponseAuthenticationResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "result": {
                        "$ref": "#/components/schemas/AuthenticationResponse"
                    },
                    "error": {
                        "type": "object"
                    }
                }
            },
            "AuthenticationResponse": {
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string"
                    },
                    "accessToken": {
                        "type": "string"
                    },
                    "authenticated": {
                        "type": "boolean"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "LogoutRequest": {
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string"
                    }
                }
            },
            "AuthenticationRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}